requirements.txt
aiohttp==3.9.1
asyncio
typing-extensions==4.8.0
python-dotenv==1.0.0
pathlib2>=2.3.7; python_version < "3.4"
openai==1.12.0
Django==5.2.3
djangorestframework==3.15.0
django-cors-headers==4.3.1
psutil==5.9.8 